requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Deprecated==1.2.13
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
numpy==1.24.3
packaging==23.1
pandocfilters==1.5.0
pluggy==1.0.0
pyasn1==0.5.0
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
wrapt==1.15.0
